Added DetailTobaccoModule, rewritten scrollViewController, bugfix and added support SwiftLint
Added DetailTobaccoModule.
Added the function to change the remote version of the database.
rewritten scrollViewController, scrollView has become private, a method has been created to define constraints, and changing the height of the content takes place inside the class.
Added SwiftLint and changed files were corrected.
BugFix: changed a bug in which relationships between database entities were not updated, removing the multi-threaded update of the database in the data manager.